Initial commit of Helios
Initial commit for Helios, a Julia testbench for the development of the
"next-gen" Carney lab IC model with an efferent gain circuit.

Initial commit includes a cleaned-up copy of the middle ear code, which
has been freed from its location within the main model loop and separated
so that it can be run in advance of the main processing loop.