Converted transmission delay to samplewise
Converted implementation of the transmission / spike-latency delay,
which is calculated based on CF by delay_cat, into a sample-wise
implementation that transmits zeros for the middle-ear output up to
a certain delay sample, and then returns actual middle-ear output
for later samples.

In testing suite, changes were made to account for this fact, and
outputs of the old model that were generated before the transmission
delay was applied are delayed post hoc to account for the difference.

Additionally, a minor adjustment was made to the unit test for
the control-path signal. Due to need to accomodate differences in delay
implementations above, we now ignore the onset response when comparing
the control-path signal for the new code to that for the old code.