Integrated all code into single loop
Removed the power-law "delaypoint" system, whereby IHC responses were
zero-padded on both sides prior to implementation of the power-law
adaptation. This removal will create some issues at onsets due to the
power law needing time to "settle in", but this could be resolved in
the future.

Because the "delaypoint" system was the last feature preventing fusion
of the entire model into a single loop, the two stages were fused in
this commit and now the entire model is implemented on a sample-wise
basis in a single loop.