Slight improvements to sim_efferent_model.m
Updated and cleaned up some of the help docs at the top of the file.

Also added some validators to the arguments block to make sure that we
are always passing row vectors and checking to make sure that parameters
with discrete numbers of options (e.g., noiseType) are checked to make
sure that the option is legitimate.