Modified Julia wrapper to load/unload library
Modified Julia wrapper to load the C library each
time we run the model and fetch a pointer to the
function and then unload the C library after running
the model. This differs from the previous approach
of implicitly loading the library (permanently for
a single Julia instance) via the ccall syntax
where the first argument is a tuple.