Added "settle duration" to C and wrapper
Added a "settle duration", currently hard coded, to the C model code
and to the MATLAB wrapper code. In the C code, during samples
below the sample determined by the settle duration of 100 ms, WDR- and
IC-driven gain control signals are fixed to zero output and the lowpass
filters do not "read in" the LSR and IC responses. The intention here
is to avoid the spurious CN/IC onsets creating strong gain changes in
response to silence. In the MATLAB code, the same duration of "settle
time" is appended to input stimuli as zeros to allow the model responses
to stabilize before stimulus-driven responses are initiated. This
additional simulation time can optionally be removed from the output
responses by `clip_settle=1`.