Updated Mex file to eliminate usage of the C model wrapper
Updated the Mex file to call `model` directly, instead of
`model_efferent_wrapper`. Parameters that are exposed in `model` but
not the wrapper (spont, CN parameters, MOC IO min/max values) are now
hardcoded in the call to model. Returns that are exposed in `model` but
not the wrapper are now allocated inside the Mex file and then freed
at the end of the file without returning to MATLAB.