Adjusted how gain fixing is implemented to make life easier
Adjusted how gain fixing is implemented in the Julia wrapper to avoid
having to maintain multiple versions or have a complicated positional
signature. Now, gain is passed as a keyword argument and the default
value is an empty vector. We check if it's still empty during evaluation
and, if so, we replace it with an appropriate zero-filled matrix.